+++ subsetter/common.py
@@ -1,6 +1,6 @@
 import logging
 import os
-from typing import Any, Dict, List, Literal, Optional, Tuple
+from typing import Any, Dict, List, Literal, Optional, Tuple, Union
 
 import sqlalchemy as sa
 from pydantic import BaseModel
@@ -13,6 +13,9 @@ DatabaseDialect = Literal[
 
 LOGGER = logging.getLogger(__name__)
 DEFAULT_DIALECT: Literal["mysql"] = "mysql"
+
+SQLKnownOperator = Literal["<", ">", "=", "<>", "!=", "like", "not like"]
+SQLLiteralType = Union[str, int, float]
 
 # pylint: disable=unused-argument
 
@@ -82,6 +85,75 @@ IsolationLevel = Literal[
 
 
 class DatabaseConfig(BaseModel):
+    class SSLConfig(BaseModel):
+        class SSLClientConfig(BaseModel):
+            cert: Optional[str] = None
+            key: Optional[str] = None
+
+        class SSLServerConfig(BaseModel):
+            rootcert: Optional[str] = None
+            verify: Literal["none", "cert", "identity"] = "none"
+
+        require: Optional[bool] = False
+        client: SSLClientConfig = SSLClientConfig()
+        server: SSLServerConfig = SSLServerConfig()
+
+        def get_connect_args(
+            self,
+            dialect: DatabaseDialect,
+            *,
+            env_prefix: Optional[str] = None,
+        ) -> Dict:
+            as_dict = self.model_dump(exclude_unset=True)
+            require = as_dict.get(
+                "require", bool(os.getenv(f"{env_prefix}SSL_REQUIRE", ""))
+            )
+            if not require:
+                return {}
+
+            server_verify = as_dict.get("server", {}).get(
+                "verify", os.getenv(f"{env_prefix}SSL_SERVER_VERIFY", "none")
+            )
+            client_cert = self.client.cert or os.getenv(f"{env_prefix}SSL_CLIENT_CERT")
+            client_key = self.client.key or os.getenv(f"{env_prefix}SSL_CLIENT_KEY")
+            server_rootcert = self.server.rootcert or os.getenv(
+                f"{env_prefix}SSL_SERVER_ROOTCERT"
+            )
+
+            result: Dict
+            if dialect == "postgres":
+                if server_verify == "identity":
+                    sslmode = "verify-full"
+                elif server_verify == "cert":
+                    sslmode = "verify-ca"
+                else:
+                    sslmode = "require"
+
+                result = {
+                    "sslcert": client_cert,
+                    "sslkey": client_key,
+                    "sslmode": sslmode,
+                    "sslrootcert": server_rootcert,
+                }
+                return {key: value for key, value in result.items() if value}
+
+            if dialect == "mysql":
+                result = {
+                    "ca": server_rootcert,
+                    "key": client_key,
+                    "cert": client_cert,
+                }
+                result = {key: value for key, value in result.items() if value}
+                if server_verify == "identity":
+                    result["verify_identity"] = True
+                elif server_verify == "cert":
+                    result["verify_cert"] = True
+                else:
+                    result["verify_cert"] = False
+                return {"ssl": result}
+
+            raise ValueError("Dialect does not support ssl")
+
     dialect: Optional[DatabaseDialect] = None
     host: Optional[str] = None
     port: Optional[int] = None
@@ -91,6 +163,7 @@ class DatabaseConfig(BaseModel):
     session_sqls: List[str] = []
     sqlite_databases: Optional[Dict[str, str]] = {}
     isolation_level: Optional[IsolationLevel] = None
+    ssl: SSLConfig = SSLConfig()
 
     def database_url(
         self,
@@ -116,10 +189,15 @@ class DatabaseConfig(BaseModel):
             isolation_level = "SERIALIZABLE"
         else:
             isolation_level = "READ COMMITTED"
+
         engine = sa.create_engine(
             self.database_url(env_prefix=env_prefix),
             isolation_level=isolation_level,
             pool_pre_ping=True,
+            connect_args=self.ssl.get_connect_args(
+                self.dialect or DEFAULT_DIALECT,
+                env_prefix=env_prefix,
+            ),
         )
 
         @sa.event.listens_for(engine, "connect")